// backend/src/send-grid/interfaces/email-temaplate.interface.ts
export interface EmailTemplate {
  to: string;
  from: string;
  subject: string;
  text: string;
  templateId: string;
  dynamicTemplateData: DynamicTemplateData;
}

export interface DynamicTemplateData {
  user: {
    name: string;
    phone: string;
    email: string;
    address: string;
  };
  products: {
    name: string;
    quantity: number;
    price: number;
  }[];
  total: number;
}